+++ NEWS.md
@@ -12,3 +12,5 @@ annolite 0.0.0.9005-6
 * The R function `dialog_default_callback()` that defined a default callback function for the bootbox dialogue has been removed. The default JavaScript callback function it defined had grown so complicated that it is unlikely that modificiations "on the fly" are possible and likely.
 * The htmlwidget included in the package has been renamed as "annotator" (previously: "fulltext").
 * The `as.fulltextdata()` has been renamed as `as.annotatordata`.
+* Calling the annotator widget is faster, resulting from an improved JavaScript implemention.
+* Argument `headline` of the `as.annotatordata()` method has now default value `NULL`, it is not necessary to provide argument.